feat: TopBarControllerV2 - selectTopBarCountV2(), selectTopBarV2() 성능 개선   1. 단순 카운팅의 경우 레디스에서 or, and 연산 후 카운팅,   2. 상품을 가져올 때는 인덱스 테이블과 1번의 결과를 and 연락해 정렬하는 기능 구현

+++ src/main/java/com/matsinger/barofishserver/domain/data/topbar/api/TopBarControllerV2.java
@@ -50,7 +50,7 @@ public class TopBarControllerV2 {
                                                                   @RequestParam(value = "filterFieldIds", required = false) String filterFieldIds) {
         CustomResponse<Long> response = new CustomResponse<>();
 
-        Long count = (long) productQueryService.countTopBarProduct(
+        Long count = productQueryService.countTopBarProduct(
                 id,
                 utils.str2IntList(filterFieldIds),
                 utils.str2IntList(categoryIds)
